Remove line chart for sent surveys and clear:  - I have removed files and code for line chart  - I have review, clear code and organize a little web files
