+++ app/src/main/assets/dashboard/css/competencies.css
