+++ app/src/main/assets/dashboard/js/translated.messages.js
@@ -21,11 +21,6 @@ var messages= {
 };
 
 /**
- * Object that holds line chart with num surveys
- */
-var surveyXMonthChart=null;
-
-/**
  * Updates every entry with the map in the given messages
  */
 var initContext = function(translated){
@@ -36,9 +31,6 @@ var initContext = function(translated){
             messages[key] = translated[key];
         }
     }
-
-    //Init SurveyXMonthChart
-    surveyXMonthChart = SentXMonthChart();
 
     updateChartTitle("multipleEventLegend",messages["multipleEventLegend"]);
 }
